Fixed how players and scores were being added and updated on the leader board.
Comments need to be added, and a tie counter needs to be implemented to ensure all 225 matches are being recorded correctly. Method being used to print the top 3 players could also be optimized.

Some players never had a name set for them so they appear on the leaderboard as "None"